+++ Tests/Runtime/VespaRuntimeTests.cs
@@ -1,0 +1,9 @@
+using NUnit.Framework;
+
+namespace LMirman.VespaIO.Tests.Runtime
+{
+	public class VespaRuntimeTests
+	{
+
+	}
+}